requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
gunicorn==23.0.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
cryptography==44.0.2
